Merge rama guti

+++ BitcoinMiner/src/main/java/UnidadDeTrabajo.java
@@ -1,18 +1,19 @@
 public class UnidadDeTrabajo {
 
-    private final int minimo;
-    private final int maximo;
+    private final long minimo;
+    private final long maximo;
     private final String texto;
 
-    public UnidadDeTrabajo(int minimo, int maximo, String texto){
+    public UnidadDeTrabajo(long minimo, long maximo, String texto){
         this.minimo = minimo;
         this.maximo = maximo;
         this.texto  = texto;
     }
-    public int getMinimo(){
+
+    public long getMinimo(){
         return this.minimo;
     }
-    public int getMaximo(){
+    public long getMaximo(){
         return this.maximo;
     }
     public String getTexto(){
